+++ src/app/core/http/api/api.service.ts
@@ -1,7 +1,5 @@
-import { StaticService } from 'src/app/config/static.service';
-
 import { Injectable } from '@angular/core';
-import { Observable, throwError } from 'rxjs';
+import { Observable, of, throwError } from 'rxjs';
 import { environment } from 'src/environments/environment';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { catchError } from 'rxjs/operators';
@@ -15,10 +13,14 @@ import {
   OfferPropertyTagResponse,
   APIToOfferShortList,
   PaginatedOfferDataFromAPI,
-  OfferFilterToAPI,
   OfferSearchFilterToAPI,
 } from './api.interfaces';
 import { AuthResponseData } from 'src/app/core/auth/auth.interfaces';
+import { LogService } from 'src/app/core/services/logger/log.service';
+import {
+  ErrorCodes,
+  ErrorHandlerService,
+} from 'src/app/core/services/error-handling/error-handling';
 
 /**
  * api.service.ts
@@ -28,11 +30,17 @@ import { AuthResponseData } from 'src/app/core/auth/auth.interfaces';
  * Updated 27.10.2021/ml
  */
 
+/* eslint-disable no-console */
+
 @Injectable({
   providedIn: 'root',
 })
 export class ApiService {
-  constructor(private http: HttpClient) {}
+  constructor(
+    private http: HttpClient,
+    private logService: LogService,
+    private errorHandler: ErrorHandlerService
+  ) {}
 
   ////////////////////////////////////////////////
   // Authenticate
@@ -46,7 +54,11 @@ export class ApiService {
         client_id: environment.clientLoginData.clientId,
         client_secret: environment.clientLoginData.clientSecret,
       })
-      .pipe(catchError(this.handleLoginError));
+      .pipe(
+        catchError((errorResponse: HttpErrorResponse) => {
+          return this.handleError(errorResponse);
+        })
+      );
   }
 
   ////////////////////////////////////////////////
@@ -82,7 +94,11 @@ export class ApiService {
         environment.apiURL + '/api/list/offer/short/paginated/' + count + '?page=' + page,
         filterSearchObj
       )
-      .pipe(catchError(this.handleError));
+      .pipe(
+        catchError((errorResponse: HttpErrorResponse) => {
+          return this.handleError(errorResponse);
+        })
+      );
   }
 
   ////////////////////////////////////////////////
@@ -90,9 +106,11 @@ export class ApiService {
   ////////////////////////////////////////////////
 
   public getAllOfferShortList(): Observable<APIToOfferShortList[]> {
-    return this.http
-      .get<APIToOfferShortList[]>(environment.apiURL + '/api/list/offer/short')
-      .pipe(catchError(this.handleError));
+    return this.http.get<APIToOfferShortList[]>(environment.apiURL + '/api/list/offer/short').pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   /**
@@ -100,45 +118,59 @@ export class ApiService {
    * @returns Observable<Offer[]>
    */
   public getAllOffers(): Observable<Offer[]> {
-    return this.http
-      .get<Offer[]>(environment.apiURL + '/api/offer')
-      .pipe(catchError(this.handleError));
+    return this.http.get<Offer[]>(environment.apiURL + '/api/offer').pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getOfferSubListWithKeyWords(keyword: string): Observable<APIToOfferShortList[]> {
-    return this.http
-      .get<Offer[]>(environment.apiURL + '/api/search/offer/sublist/' + keyword)
-      .pipe(catchError(this.handleError));
+    return this.http.get<Offer[]>(environment.apiURL + '/api/search/offer/sublist/' + keyword).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getOfferLatest(): Observable<APIToOfferShortList[]> {
-    return this.http
-      .get<Offer[]>(environment.apiURL + '/api/search/offer/latest')
-      .pipe(catchError(this.handleError));
+    return this.http.get<Offer[]>(environment.apiURL + '/api/search/offer/latest').pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getOffer(id: number): Observable<Offer> {
-    return this.http
-      .get<Offer>(environment.apiURL + '/api/offer/' + id)
-      .pipe(catchError(this.handleError));
+    return this.http.get<Offer>(environment.apiURL + '/api/offer/' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public putOffer(id: number, data: OfferToAPI): Observable<Offer> {
-    return this.http
-      .put<Offer>(environment.apiURL + '/api/offer/' + id, data)
-      .pipe(catchError(this.handleError));
+    return this.http.put<Offer>(environment.apiURL + '/api/offer/' + id, data).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public postOffer(data: OfferToAPI): Observable<Offer> {
-    return this.http
-      .post<Offer>(environment.apiURL + '/api/offer', data)
-      .pipe(catchError(this.handleError));
+    return this.http.post<Offer>(environment.apiURL + '/api/offer', data).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public deleteOffer(id: number) {
-    return this.http
-      .delete(environment.apiURL + '/api/offer/' + id)
-      .pipe(catchError(this.handleError));
+    return this.http.delete(environment.apiURL + '/api/offer/' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   ////////////////////////////////////////////////
@@ -147,48 +179,65 @@ export class ApiService {
   ////////////////////////////////////////////////
 
   public getAllSubscriptions(): Observable<SubscriptionData[]> {
-    return this.http
-      .get<SubscriptionData[]>(environment.apiURL + '/api/subscription')
-      .pipe(catchError(this.handleError));
+    return this.http.get<SubscriptionData[]>(environment.apiURL + '/api/subscription').pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getSubscriptionById(id: number): Observable<SubscriptionData> {
-    return this.http
-      .get<SubscriptionData>(environment.apiURL + '/api/subscription/' + id)
-      .pipe(catchError(this.handleError));
+    return this.http.get<SubscriptionData>(environment.apiURL + '/api/subscription/' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getSubscriptionByIds(user_id: number, offer_id: number): Observable<SubscriptionData[]> {
     return this.http
       .get<SubscriptionData[]>(environment.apiURL + '/api/subscription/' + user_id + '/' + offer_id)
-      .pipe(catchError(this.handleError));
+      .pipe(
+        catchError((errorResponse: HttpErrorResponse) => {
+          return this.handleError(errorResponse);
+        })
+      );
   }
 
   public getUserSubscriptions(user_id: number) {
     return this.http
       .get<SubscriptionData[]>(environment.apiURL + '/api/subscription/user/' + user_id)
-      .pipe(catchError(this.handleError));
+      .pipe(
+        catchError((errorResponse: HttpErrorResponse) => {
+          return this.handleError(errorResponse);
+        })
+      );
   }
 
   public putSubscription(id: number, data: Object): Observable<SubscriptionData> {
-    console.log(data);
     return this.http
       .put<SubscriptionData>(environment.apiURL + '/api/subscription/' + id, data)
-      .pipe(catchError(this.handleError));
+      .pipe(
+        catchError((errorResponse: HttpErrorResponse) => {
+          return this.handleError(errorResponse);
+        })
+      );
   }
 
   public postSubscription(data: Object): Observable<SubscriptionData> {
-    console.log(data);
-    return this.http
-      .post<SubscriptionData>(environment.apiURL + '/api/subscription', data)
-      .pipe(catchError(this.handleError));
+    return this.http.post<SubscriptionData>(environment.apiURL + '/api/subscription', data).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public deleteSubscription(id: number) {
-    console.log('del Subscription' + id);
-    return this.http
-      .delete(environment.apiURL + '/api/subscription/' + id)
-      .pipe(catchError(this.handleError));
+    return this.http.delete(environment.apiURL + '/api/subscription/' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   ////////////////////////////////////////////////
@@ -196,39 +245,51 @@ export class ApiService {
   ////////////////////////////////////////////////
 
   public getAllUsers(): Observable<User[]> {
-    return this.http
-      .get<User[]>(environment.apiURL + '/api/user')
-      .pipe(catchError(this.handleError));
+    return this.http.get<User[]>(environment.apiURL + '/api/user').pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getUser(id: number): Observable<User> {
-    return this.http
-      .get<User>(environment.apiURL + '/api/user/' + id)
-      .pipe(catchError(this.handleError));
+    return this.http.get<User>(environment.apiURL + '/api/user/' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getUserByEmail(email: string): Observable<User> {
-    return this.http
-      .get<User>(environment.apiURL + '/api/user/email?email=' + email)
-      .pipe(catchError(this.handleError));
+    return this.http.get<User>(environment.apiURL + '/api/user/email?email=' + email).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public putUser(id: number, data: Object): Observable<User> {
-    return this.http
-      .put<User>(environment.apiURL + '/api/user/' + id, data)
-      .pipe(catchError(this.handleError));
+    return this.http.put<User>(environment.apiURL + '/api/user/' + id, data).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public postUser(data: Object): Observable<User> {
-    return this.http
-      .post<User>(environment.apiURL + '/api/user', data)
-      .pipe(catchError(this.handleError));
+    return this.http.post<User>(environment.apiURL + '/api/user', data).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public deleteUser(id: number) {
-    return this.http
-      .delete(environment.apiURL + '/api/user/' + id)
-      .pipe(catchError(this.handleError));
+    return this.http.delete(environment.apiURL + '/api/user/' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   ////////////////////////////////////////////////
@@ -236,15 +297,19 @@ export class ApiService {
   ////////////////////////////////////////////////
 
   public getInstitutions(): Observable<Institution[]> {
-    return this.http
-      .get<Institution[]>(environment.apiURL + '/api/institution/')
-      .pipe(catchError(this.handleError));
+    return this.http.get<Institution[]>(environment.apiURL + '/api/institution/').pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public getInstitutionById(id: number): Observable<Institution> {
-    return this.http
-      .get<Institution>(environment.apiURL + '/api/institution/' + id)
-      .pipe(catchError(this.handleError));
+    return this.http.get<Institution>(environment.apiURL + '/api/institution/' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public postInstitution(institution: Institution): Observable<Institution> {
@@ -252,9 +317,11 @@ export class ApiService {
       title: institution.title,
       url: institution.url,
     };
-    return this.http
-      .post<Institution>(environment.apiURL + '/api/institution/', data)
-      .pipe(catchError(this.handleError));
+    return this.http.post<Institution>(environment.apiURL + '/api/institution/', data).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   public putInstitution(institution: Institution): Observable<Institution> {
@@ -263,17 +330,24 @@ export class ApiService {
       title: institution.title,
       url: institution.url,
     };
-    return this.http
-      .put<Institution>(environment.apiURL + '/api/institution/' + id, data)
-      .pipe(catchError(this.handleError));
+    return this.http.put<Institution>(environment.apiURL + '/api/institution/' + id, data).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
-  public deleteInstitution(institution: Institution) {
+  public deleteInstitution(institution: Institution): Observable<any> {
     let id = institution.id;
-    if (id == undefined) return;
-    return this.http
-      .delete(environment.apiURL + '/api/institution' + id)
-      .pipe(catchError(this.handleError));
+    if (id == undefined) {
+      // TODO: Rückgabewert nochmal checken
+      return of(false);
+    }
+    return this.http.delete(environment.apiURL + '/api/institution' + id).pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   ////////////////////////////////////////////////
@@ -281,9 +355,11 @@ export class ApiService {
   ///////////////////////////////////////////////
 
   public getFilterTags(): Observable<OfferPropertyTagResponse> {
-    return this.http
-      .get<OfferPropertyTagResponse>(environment.apiURL + '/api/filter/tags')
-      .pipe(catchError(this.handleError));
+    return this.http.get<OfferPropertyTagResponse>(environment.apiURL + '/api/filter/tags').pipe(
+      catchError((errorResponse: HttpErrorResponse) => {
+        return this.handleError(errorResponse);
+      })
+    );
   }
 
   // Die Filter sind im Moment identisch mit den Properties
@@ -295,52 +371,38 @@ export class ApiService {
   // ErrorHandling
   ////////////////////////////////////////////////
   // Todo: Fehlermeldungstexte für die GUI
-  private handleError(errorRes: HttpErrorResponse) {
-    let errorMessage = 'Ein unbekannter Fehler ist aufgetreten.';
-    const e401 = 'Fehler beim Zugriff auf Benutzerdaten.';
-    const e404 = 'Es wurden keine Daten gefunden.';
-    const e422 = 'Fehler beim Speichern der Daten.';
-    const e500 = 'Ein Serverfehler ist aufgetreten.';
-    console.log('ErrorHandler : ', errorRes);
+  private handleError(errorRes: HttpErrorResponse): Observable<never> {
+    let errorCode = ErrorCodes.UNKNOWN;
 
-    // Mehrere Möglichkeiten
-    // Abfrage auf Status
-    // 401 - Unauthorized
     switch (errorRes.status) {
+      case 0:
+        errorCode = ErrorCodes.E0;
+        break;
+      case 400:
+        errorCode = ErrorCodes.E400;
+        break;
       case 401:
-        errorMessage = e401;
-        return throwError(errorMessage);
+        errorCode = ErrorCodes.E401;
+        break;
+      case 403:
+        errorCode = ErrorCodes.E403;
+        break;
       case 404:
-        errorMessage = e404;
-        return throwError(errorMessage);
+        errorCode = ErrorCodes.E404;
+        break;
       case 422:
-        errorMessage = e422;
-        return throwError(errorMessage);
+        errorCode = ErrorCodes.E422;
+        break;
+      case 429:
+        errorCode = ErrorCodes.E429;
+        break;
       case 500:
-        errorMessage = e500;
-        return throwError(errorMessage);
+        errorCode = ErrorCodes.E500;
+        break;
     }
+    let newError = new Error(errorCode);
 
-    // oder Abfrage der Message
-    if (!errorRes.error) {
-      return throwError(errorMessage);
-    } else {
-      // return an observable with a user-facing error message
-      console.log('ErrorMessage: ', errorRes.error.message);
-      switch (errorRes.error.message) {
-        case 'Unauthenticated.':
-          errorMessage = e401;
-          break;
-      }
-    }
-
-    return throwError(errorMessage);
-  }
-
-  private handleLoginError(errorRes: HttpErrorResponse) {
-    // Todo: Fehlermeldungstexte für die GUI
-    let errorMessage = 'Fehler beim Login.';
-    console.log('LoginError : ', errorRes);
-    return throwError(errorMessage);
+    console.log('ErrorResponse:', errorRes);
+    return throwError(() => newError);
   }
 }