+++ src/app/core/interceptors/token-check.interceptor.ts
@@ -1,9 +1,16 @@
 import { Router } from '@angular/router';
 import { AuthService } from './../auth/auth.service';
 import { Injectable } from '@angular/core';
-import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
-import { Observable, EMPTY } from 'rxjs';
-import { take, exhaustMap, map } from 'rxjs/operators';
+import {
+  HttpRequest,
+  HttpHandler,
+  HttpEvent,
+  HttpInterceptor,
+  HttpResponse,
+  HttpErrorResponse,
+} from '@angular/common/http';
+import { Observable, EMPTY, throwError } from 'rxjs';
+import { take, exhaustMap, map, catchError } from 'rxjs/operators';
 import { StaticService } from 'src/app/config/static.service';
 import { UserData } from './../data/user/user-data.interface';
 import { TokenService } from 'src/app/core/services/token-check/token.service';
@@ -28,10 +35,8 @@ export class TokenCheckInterceptor implements HttpInterceptor {
           const isValid = this.tokenService.isTokenValid(token);
 
           if (isValid) {
-            console.log('Token check (Interceptor) ok');
             return next.handle(request);
           } else {
-            console.log('Token expired');
             // Erstmal kein Logout
             return next.handle(request);
             // this.handleExpiredToken();
@@ -50,5 +55,20 @@ export class TokenCheckInterceptor implements HttpInterceptor {
     this.authService.logoutUserOnTokenExpired().subscribe((value) => {
       this.router.navigate([this.lnkLogin]);
     });
+  }
+
+  private checkResponseIfLoggedIn(request: HttpRequest<unknown>, next: HttpHandler) {
+    return next.handle(request).pipe(
+      map((event: HttpEvent<any>) => {
+        if (event instanceof HttpResponse) {
+          //console.log('event--->>>', event);
+        }
+        return event;
+      }),
+      catchError((errorResponse: HttpErrorResponse) => {
+        //console.log('ErrorResponse', errorResponse);
+        return throwError(() => new Error('Login Error in Interceptor'));
+      })
+    );
   }
 }